test updated numpy build with cross-compilation support, run on ubuntu-22.04

+++ numpy/cppbuild.sh
@@ -113,11 +113,19 @@ case $PLATFORM in
         arm-linux-gnueabihf-strip $(find ../ -iname *.so)
         ;;
     linux-arm64)
-        rm -f meson.build pyproject.toml
-        ls -l
-        mv pyproject.toml.setuppy pyproject.toml
-        ATLAS=None CC="aarch64-linux-gnu-gcc -mabi=lp64" CFLAGS="-O2" "$PYTHON_BIN_PATH" -m pip install . --prefix $INSTALL_PATH
-        aarch64-linux-gnu-strip $(find ../ -iname *.so)
+        HOST_ARCH="$(uname -m)"
+        if [[ $HOST_ARCH == *"arm"* ]]; then
+          echo "Detected arm arch so not cross compiling";
+          ATLAS=None CC="gcc -std=c99" "$PYTHON_BIN_PATH" -m pip install . --prefix $INSTALL_PATH
+          strip $(find ../ -iname *.so)
+        else
+          rm -f meson.build pyproject.toml
+          ls -l
+          mv pyproject.toml.setuppy pyproject.toml
+          ATLAS=None CC="aarch64-linux-gnu-gcc -mabi=lp64" CFLAGS="-O2" "$PYTHON_BIN_PATH" -m pip install . --prefix $INSTALL_PATH
+          aarch64-linux-gnu-strip $(find ../ -iname *.so)
+          echo "Detected non arm arch so cross compiling";
+        fi
         ;;
     linux-ppc64le)
         ATLAS=None CC="powerpc64le-linux-gnu-gcc -m64" "$PYTHON_BIN_PATH" -m pip install . --prefix $INSTALL_PATH
